btrfs: reorder arguments so that options come first
The option parser in the btrfs utility is going to be updated and will
accept arguments in a stricter form, namely that options must come
before their non-option argument. Otherwise eg. the -f option in send
would be understood as another path and not an option leading to many
test failures.

The canonical form should be:

  btrfs command subcommand [-options] [arguments]

Signed-off-by: David Sterba <dsterba@suse.com>