fstest: Crashmonkey rename tests ported to xfstest
This patch aims to add tests to the xfstest suite to check whether
renamed files recover after a crash. These test cases were generated by
CrashMonkey, a crash-consistency testing framework built at the SASLab
at UT Austin.

This patch batches 37 tests into a single test, each of which creates a
file, renames it, syncs one of the files/parent directories, and ensures
that the synced file/dir is equivalent before and after a crash.

This test takes around 12-15 seconds to run locally and is thus placed
in the 'quick' group.

Signed-off-by: Arvind Raghavan <raghavan.arvind@gmail.com>